@@ -23,11 +23,11 @@
         }
     },
     "require": {
-        "php": "^7.1",
+        "php": "^7.2",
         "psr/log": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0|^8.0|^9.0",
+        "phpunit/phpunit": "^8.0|^9.0",
         "php-coveralls/php-coveralls": "^2.0",
         "squizlabs/php_codesniffer": "^3.5"
     }
